Move the SentryOptions back into Platforms/Android
