+++ src/Sentry/Platforms/Android/SentryOptions.cs
@@ -11,6 +11,20 @@ public partial class SentryOptions
     public AndroidOptions Android { get; }
 
     /// <summary>
+    /// Gets or sets whether when LogCat logs are attached to events.
+    /// The default is <see cref="LogCatIntegrationType.None"/>
+    /// </summary>
+    /// <seealso cref="LogCatMaxLines" />
+    public LogCatIntegrationType LogCatIntegration { get; set; } = LogCatIntegrationType.None;
+
+    /// <summary>
+    /// Gets or sets the maximum number of lines to read from LogCat logs.
+    /// The default value is 1000.
+    /// </summary>
+    /// <seealso cref="LogCatIntegration" />
+    public int LogCatMaxLines { get; set; } = 1000;
+
+    /// <summary>
     /// Provides additional options for the Android platform.
     /// </summary>
     public class AndroidOptions
@@ -232,7 +246,8 @@ public partial class SentryOptions
         /// <example>
         /// 'java.util.customcode.', 'io.sentry.samples.'
         /// </example>
-        public void AddInAppInclude(string prefix){
+        public void AddInAppInclude(string prefix)
+        {
             InAppIncludes ??= new List<string>();
             InAppIncludes.Add(prefix);
         }
