+++ src/Sentry/SentryOptions.cs
@@ -967,22 +967,6 @@ public class SentryOptions
     /// </summary>
     public Func<bool>? CrashedLastRun { get; set; }
 
-#if ANDROID
-    /// <summary>
-    /// Gets or sets whether when LogCat logs are attached to events.
-    /// The default is <see cref="LogCatIntegrationType.None"/>
-    /// </summary>
-    /// <seealso cref="LogCatMaxLines" />
-    public LogCatIntegrationType LogCatIntegration { get; set; } = LogCatIntegrationType.None;
-
-    /// <summary>
-    /// Gets or sets the maximum number of lines to read from LogCat logs.
-    /// The default value is 1000.
-    /// </summary>
-    /// <seealso cref="LogCatIntegration" />
-    public int LogCatMaxLines { get; set; } = 1000;
-#endif
-
     /// <summary>
     /// <para>
     ///     Gets the <see cref="Instrumenter"/> used to create spans.
